Added Piral to list
I'd potentially also add a section that featurevisor could be consumed from a microfrontend discovery service, too. This way, you would not need to consume it within a microfrontend (but, of course, for whatever reason one might still do that). In many cases just disabling or enabling the microfrontend for certain users might be sufficient and the most consistent way of dealing with functionality. (One of many examples: https://youtu.be/Oc2ohxEpUpA?t=2176)